Retain fact code in files even when not checking on load
When running something like eastwood to lint test files, it's nice to be
able to not also run tests while linting.  Unfortunately setting
include-midje-checks to false means the test code that's meant to be
linted macro-expands to nil, so this patch puts the expanded code into a
defn where it won't be run but can still be linted.