Move determination of :colorize [Fixes #458]
The environment variable MIDJE_COLORIZE is now read once, and put into the
config. From there on out, it is treated just like any of the other
configuration flags.

Normalization of the :colorize config key was moved from the
midje.emission.colorize namespace to the midje.config namespace.

Previous order of precedence was (increasing order):
- Config file
- MIDJE_COLORIZE

Previous order of precedence was (increasing order):
- Config file
- MIDJE_COLORIZE
- Any programmatic changes to the config (e.g. via with-augmented-config)