Improving control of colorize (#459)
* Check colorization config on every call

Issue #458.

* Change test to reflect reality

Issue #458.

The test was passing before, but only because the `:times 0` was not being
honored by `preprequiste`. Now the test reflects reality, but we'll want to
modify implementation so that we change change `:times 1` back to `0` and still
have them pass.

* Move determination of :colorize [Fixes #458]

The environment variable MIDJE_COLORIZE is now read once, and put into the
config. From there on out, it is treated just like any of the other
configuration flags.

Normalization of the :colorize config key was moved from the
midje.emission.colorize namespace to the midje.config namespace.

Previous order of precedence was (increasing order):
- Config file
- MIDJE_COLORIZE

Previous order of precedence was (increasing order):
- Config file
- MIDJE_COLORIZE
- Any programmatic changes to the config (e.g. via with-augmented-config)

* Fix arity breakage on colorizing functions

Issue #458.

The colorizing functions should take any number of arguments (like str), but I
accidentally made them only accept one.

* Refactor colorize to not use a macro

Issue #458.

* Reset config after modification test

Issue #458.